chore(tailwind): extract selector processing from sanitization functions
